:art: Add styles to button and use ripple effect
The theme is necessary to obtain this ripple effect. If we want to get rid of the ripple on the padding, a custom ripple drawable should be used with inset. Moreover, the foreground needs to be used which causes some phones with API < to be unsupported. Maybe use a transparent background ?